requirements.txt
arxiv
pyzotero
requests
pytz
python-dotenv
asyncio
aiohttp
pyyaml